// src/QuizFE/ClientApp/src/app/models/Registration.ts
export class UserDataModel { 
  email: string = '';
  role: string = ''; 
}

export class UserRegistration {  
  teamName: string;
  email: string;
  email2: string;
  email3: string; 
  contestantName: string;
  contestantName2: string;
  contestantName3: string;
  phone: string;
  phone2: string;
  phone3: string;
  contact: string;
  quizName: string;
  quizType: string
}

export class SignUp {
  email: string;
  password: string;
  role: string;
  status: string="Activation Pending";
  source: string;
  statusCode: number;
  url: string = "";
  firstname: string = "";
  lastname: string = "";
}
export class QuizAdv {
  quizName: string;
  quizType: string;
  teamSize: number;
  teamName: string;
}